requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
pip==23.0.1
python-dotenv==1.0.0
requests==2.28.2
setuptools==67.6.0
urllib3==1.26.15
wheel==0.40.0
